requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
